Revert "refactor: 인수 테스트로 변경"
This reverts commit 3424566e0ab6d96724916607aa1ec362f2be533b.
