+++ src/test/java/roomescape/reservation/controller/ReservationTimeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.acceptance;
+package roomescape.reservation.controller;
 
 import static org.hamcrest.Matchers.is;
 
@@ -6,20 +6,15 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.util.HashMap;
 import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation.repository.ReservationTimeRepository;
-import roomescape.util.AcceptanceTest;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
 
-class ReservationTimeAcceptanceTest extends AcceptanceTest {
-    @Autowired
-    ReservationTimeRepository reservationTimeRepository;
-
-    @Autowired
-    ReservationRepository reservationRepository;
-
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+class ReservationTimeControllerTest {
     @DisplayName("시간 생성 시, 200을 반환한다.")
     @Test
     void create() {
@@ -71,6 +66,7 @@ class ReservationTimeAcceptanceTest extends AcceptanceTest {
                 .when().post("/times")
                 .then().log().all()
                 .statusCode(200);
+
 
         //when & then
         RestAssured.given().log().all()
